+++ .dockerignore
@@ -6,4 +6,3 @@ infra
 data/prep
 data/raw
 prep/stage
-diagram.png
