+++ .streamlit/config.toml
@@ -102,7 +102,7 @@ port = 8080
 
 # The base path for the URL where Streamlit should be served from.
 # Default: ""
-baseUrlPath = "/"
+baseUrlPath = ""
 
 # Enables support for Cross-Origin Request Sharing (CORS) protection, for added security.
 # Due to conflicts between CORS and XSRF, if `server.enableXsrfProtection` is on and `server.enableCORS` is off at the same time, we will prioritize `server.enableXsrfProtection`.
