--- apprunner.yaml
@@ -1,12 +1,0 @@
-version: 1.0
-runtime: python3
-build:
-  commands:
-    build:
-    - pip install poetry
-    - poetry config virtualenvs.create false
-    - poetry install --no-dev
-run:
-  command: make streamlit_local
-  network:
-    port: 8000
