--- heroku.yml
@@ -1,5 +1,0 @@
-build:
-  docker:
-    web: Dockerfile
-run:
-  web: make streamlit_cloud
